requirements.txt
numpy
pandas
dlib
face_recognition
opencv-python
uvicorn
google-cloud-storage
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
python-dotenv
Jinja2==3.0.3
python-multipart
Pillow
matplotlib

